contract-classifier/requirements.txt
langchain
langgraph
langchain-community
langchain-groq
docling
pydantic
python-dotenv

